#27 - moves home screen tests to home module
